Made changes to how arrows are displayed to allow better arrow skin customisation
Arrows positions are now defined by the center of the texture instead of a corner, so you will be able to make arrows that will not have its resolution forced to 150x150. This wil allow arrows with glow effects without affecting the size of the arrows. Maybe better arrow press animations in 1.9?